+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## v0.8.3
 
 * ⭐ Neu: Ausgabe an EEP-Web enthält nun auch Fehlermeldungen die mit assert ausgegeben werden.
+* ⭐ Neu: Der Name wird nun auch für Speicherplätze angezeigt, die mit AkSpeicherHilfe.registriereId(id, name) angelegt 
+  wurden
 
 ## v0.8.2